+++ pkg/agent/collectors/logs/configmap.go
@@ -49,7 +49,7 @@ func ConfigMap(cluster *clusterv1.Cluster) *v1.ConfigMap {
 	}
 }
 
-func (s *Service) GenerateAlloyLogsConfigMapData(ctx context.Context, cluster *clusterv1.Cluster, observabilityBundleVersion semver.Version, networkMonitoringEnabled bool) (map[string]string, error) {
+func (s *Service) GenerateAlloyLogsConfigMapData(ctx context.Context, cluster *clusterv1.Cluster, observabilityBundleVersion semver.Version, loggingEnabled bool, networkMonitoringEnabled bool) (map[string]string, error) {
 	// Get tenant IDs
 	tenants, err := s.TenantRepository.List(ctx)
 	if err != nil {
@@ -84,6 +84,7 @@ func (s *Service) GenerateAlloyLogsConfigMapData(ctx context.Context, cluster *c
 		provider,
 		s.Config.Cluster.InsecureCA,
 		enableNodeFiltering,
+		loggingEnabled,
 		networkMonitoringEnabled,
 		s.Config.Cluster,
 	)
@@ -106,11 +107,12 @@ func generateAlloyLoggingConfig(
 	provider string,
 	insecureCA bool,
 	enableNodeFiltering bool,
+	enableLogging bool,
 	enableNetworkMonitoring bool,
 	clusterConfig config.ClusterConfig,
 ) (string, error) {
 	// Generate River configuration
-	alloyConfig, err := generateAlloyConfig(tenants, cluster, org, provider, insecureCA, enableNodeFiltering, enableNetworkMonitoring, clusterConfig)
+	alloyConfig, err := generateAlloyConfig(tenants, cluster, org, provider, insecureCA, enableNodeFiltering, enableLogging, enableNetworkMonitoring, clusterConfig)
 	if err != nil {
 		return "", err
 	}
@@ -123,6 +125,7 @@ func generateAlloyLoggingConfig(
 		AlloyImageTag                    *string
 		DefaultWorkloadClusterNamespaces []string
 		DefaultWriteTenant               string
+		LoggingEnabled                   bool
 		NetworkMonitoringEnabled         bool
 		NodeFilteringEnabled             bool
 		IsWorkloadCluster                bool
@@ -131,6 +134,7 @@ func generateAlloyLoggingConfig(
 		AlloyConfig:                      alloyConfig,
 		DefaultWorkloadClusterNamespaces: defaultWorkloadClusterNamespaces,
 		DefaultWriteTenant:               organization.GiantSwarmDefaultTenant,
+		LoggingEnabled:                   enableLogging,
 		NetworkMonitoringEnabled:         enableNetworkMonitoring,
 		NodeFilteringEnabled:             enableNodeFiltering,
 		IsWorkloadCluster:                isWorkloadCluster,
@@ -161,6 +165,7 @@ func generateAlloyConfig(
 	provider string,
 	insecureCA bool,
 	enableNodeFiltering bool,
+	enableLogging bool,
 	enableNetworkMonitoring bool,
 	clusterConfig config.ClusterConfig,
 ) (string, error) {
@@ -182,6 +187,7 @@ func generateAlloyConfig(
 		RemoteTimeout            string
 		IsWorkloadCluster        bool
 		NodeFilteringEnabled     bool
+		LoggingEnabled           bool
 		NetworkMonitoringEnabled bool
 		InsecureSkipVerify       bool
 		SecretName               string
@@ -201,6 +207,7 @@ func generateAlloyConfig(
 		RemoteTimeout:            commonmonitoring.LokiRemoteTimeout,
 		IsWorkloadCluster:        isWorkloadCluster,
 		NodeFilteringEnabled:     enableNodeFiltering,
+		LoggingEnabled:           enableLogging,
 		NetworkMonitoringEnabled: enableNetworkMonitoring,
 		InsecureSkipVerify:       insecureCA,
 		SecretName:               apps.AlloyLogsAppName,